+++ src/main/java/ma/vi/esql/exec/function/date/EndOfYear.java
@@ -20,38 +20,28 @@ import static java.util.Collections.singletonList;
 import static ma.vi.esql.translation.Translatable.Target.*;
 
 /**
- * Function returning the start of the hour of the supplied date.
+ * Function to get the end of the year of a date.
  *
  * @author Vikash Madhow (vikash.madhow@gmail.com)
  */
-public class StartOfHour extends Function {
-  public StartOfHour() {
-    super("startofhour", Types.DateType,
+public class EndOfYear extends Function {
+  public EndOfYear() {
+    super("endofmonth", Types.DateType,
           singletonList(new FunctionParam("s", Types.DateType)));
   }
 
   @Override
-  public String translate(FunctionCall         call,
-                          Target               target,
-                          EsqlConnection       esqlCon,
-                          EsqlPath             path,
-                          PMap<String, Object> parameters,
-                          Environment          env) {
+  public String translate(FunctionCall call, Target target, EsqlConnection esqlCon, EsqlPath path, PMap<String, Object> parameters, Environment env) {
     List<Expression<?, ?>> args = call.arguments();
     String arg = args.get(0).translate(target, esqlCon, path.add(args.get(0)), env).toString();
     if (target == POSTGRESQL) {
-      return "date_trunc('hour', " + arg + ")";
+      return "(date_trunc('year', " + arg + ") + interval '1 year - 1 day')::date";
 
     } else if (target == SQLSERVER) {
-      // return "datetrunc(hour, " + arg + ")";
-      return "datetime2fromparts(year("  + arg + "), "
-                              + "month(" + arg + "), "
-                              + "day("   + arg + "), "
-                              + "datepart(hour, "  + arg + "), "
-                              + "0, 0, 0, 0)";
+      return "datefromparts(year(" + arg + "), 12, 31)";
 
     } else if (target == JAVASCRIPT) {
-      return "new Date(" + arg + ".setHours(" + arg + ".getHours(), 0, 0, 0))";
+      return "new Date(" + arg + ".getFullYear(), 11, 31)";
 
     } else {
       return name + '(' + arg + ')';